Save out the code to the "out_list" in only one place:
I was saving the code to the out_list separately for the CIC, CRS, and CRS_pw models: once in section 2. LEAD 210 MODEL, for the model itself, and once in the section 4. AGE DEPTH MODEL, for the interpolated model, after writing the .txt file output.
For consistency, I am now saving the code systematically after writing it up as a txt file. This is already what I was doing for CFCS.

At the moment, I did not remove the code from the function, I just commented it. I think having it here can serve me as note of the way I called each part of the model.