change in the outlist of the max depth when sed change
In the output list that can be accessed when creating an object, I was saving for each section the min and max depth of the section. However, the sediment change is given as only one depth. So if a sediment change occured at 165mm according to the user, and the sampling resoluton was 10 mm, the first sed change would have been for the depths 0-170 and the second for the depths 160-end of the core.

The output files are correct, as well as what was printed in the console, this only affected the output table, because I was building up from the first sed change.

Now, the new table will read sediment change for depth from 0-165 and 165-end of core.